Add ShipSunkOrNotProbability module to refactor the main orchestrator method
